+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Renamed `logtesting.AssertLogEvents` to `AssertLogEventsEqual`
+- Renamed `AssertLogEventsEqual` to `LogEventsEqual`
 
 ### Removed